+++ Changelog.md
@@ -1,5 +1,11 @@
 # NEventStore.Persistence.MongoDB
 
+## vNext
+
+### Breaking Changes
+
+- Removed Virtual Methods. Prefer encapsulation over inheritance.
+
 ## 6.0.0
 
 __Version 6.x is not backwards compatible with version 5.x.__ Updating to NEventStore 6.x without doing some preparation work will result in problems.